implemented the core scan for engine

+++ src/main/java/com/buxxy/buxxy_fraud_engine/model/Device.java
@@ -18,7 +18,7 @@ public class Device {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
-    private int deviceId;
+    private long deviceId;
 
     private Long userId;
